feat: the user should database type
