+++ packages/backend/.env
@@ -9,11 +9,7 @@ AP_SANDBOX_RUN_TIME_SECONDS=120
 AP_EXECUTION_MODE=UNSANDBOXED
 
 AP_TRIGGER_DEFAULT_POLL_INTERVAL=1
-AP_POSTGRES_DATABASE="activepieces"
-AP_POSTGRES_HOST="localhost"
-AP_POSTGRES_PASSWORD="A79Vm5D4p2VQHOp2gd5"
-AP_POSTGRES_PORT="5432"
-AP_POSTGRES_USERNAME="postgres"
+AP_DB_TYPE="SQLITE3"
 AP_QUEUE_MODE="MEMORY"
 
 AP_TELEMETRY_ENABLED="false"
