+++ packages/backend/src/app/app-event-routing/app-event-routing.entity.ts
@@ -1,5 +1,5 @@
 import { FlowId, ProjectId } from '@activepieces/shared'
-import { ApIdSchema, BaseColumnSchemaPart } from '../helper/base-entity'
+import { ApIdSchema, BaseColumnSchemaPart } from '../helper/database-common'
 import { EntitySchema } from 'typeorm'
 
 export type AppEventRoutingId = string
