+++ packages/backend/src/app/flows/flow-instance/flow-instance.entity.ts
@@ -1,6 +1,6 @@
 import { EntitySchema } from 'typeorm'
 import { Flow, FlowInstance, FlowVersion } from '@activepieces/shared'
-import { ApIdSchema, BaseColumnSchemaPart, JSONB_COLUMN_TYPE } from '../../helper/base-entity'
+import { ApIdSchema, BaseColumnSchemaPart, JSONB_COLUMN_TYPE } from '../../helper/database-common'
 
 export type FlowInstanceSchema = {
     flow: Flow
