+++ packages/backend/src/app/helper/system/system-prop.ts
@@ -10,6 +10,7 @@ export enum SystemProp {
     POSTGRES_HOST = 'POSTGRES_HOST',
     LICENSE_KEY = 'LICENSE_KEY',
     ENCRYPTION_KEY = 'ENCRYPTION_KEY',
+    DB_TYPE = 'DB_TYPE',
     POSTGRES_PASSWORD = 'POSTGRES_PASSWORD',
     POSTGRES_PORT = 'POSTGRES_PORT',
     POSTGRES_SSL_CA = 'POSTGRES_SSL_CA',
