+++ packages/backend/src/app/store-entry/store-entry-entity.ts
@@ -1,5 +1,5 @@
 import { EntitySchema } from 'typeorm'
-import { ApIdSchema, BaseColumnSchemaPart, JSONB_COLUMN_TYPE } from '../helper/base-entity'
+import { ApIdSchema, BaseColumnSchemaPart, JSONB_COLUMN_TYPE } from '../helper/database-common'
 import { StoreEntry } from '@activepieces/shared'
 
 type StoreEntrySchema = StoreEntry
