+++ packages/backend/src/app/webhooks/webhook-simulation/webhook-simulation-entity.ts
@@ -1,6 +1,6 @@
 import { EntitySchema } from 'typeorm'
 import { WebhookSimulation } from '@activepieces/shared'
-import { ApIdSchema, BaseColumnSchemaPart } from '../../helper/base-entity'
+import { ApIdSchema, BaseColumnSchemaPart } from '../../helper/database-common'
 
 export type WebhookSimulationSchema = WebhookSimulation
